+++ internal/bincache/exec_child_windows.go
@@ -8,10 +8,14 @@ import (
 	"os/exec"
 )
 
-func execChild(bin string, args, env []string) error {
+func execChild(bin string, args, env []string, afterChild func()) error {
 	cmd := exec.Command(bin, args...)
 	cmd.Stdin, cmd.Stdout, cmd.Stderr, cmd.Env = os.Stdin, os.Stdout, os.Stderr, env
-	if err := cmd.Run(); err != nil {
+	err := cmd.Run()
+	if afterChild != nil {
+		afterChild()
+	}
+	if err != nil {
 		var exitErr *exec.ExitError
 		if errors.As(err, &exitErr) {
 			os.Exit(exitErr.ExitCode()) //nolint:forbidigo // foreground wrapper preserves the pipeline's exit status